Write some tests for infrastructure code
It's a little unreliable at times, so good to find out why.

For each test, we spin up a container and copy the tunnel over to it. We
then deploy some rules using this tunnel and check the response. It's
not perfect - very much black-box testing - but useful to do.

 - Fix crash in systemd scraper caused by treating the journal as an
   OCaml value.

 - apply now returns the number of changed and failed keys, instead of
   exiting the process.