Try to separate keys and their values
Before we defined resource as (R, key, value). We now just define it as
(R, key) and have a separate mapping of (R, key) -> value. This should
make it easier to add alternative key types in the future.

This is definitely not perfect and some of the interface has got much
uglier. However, it's a useful first step before we commit to a better
interface.