Clean up log handling a little bit
Our log reporter needs to run in an Eio context as it uses Eio
Fiber-local variables. Thus we need to instantiate it slightly later
than normal.

I'll be the first to admit this is not nice, but not sure how else
better to handle this! I think in the future we may want to gate reading
this variable (i.e. only do so on the main thread and once a global flag
has been set).