Treat executors as a resource dependency
Previously we created the executor after building dependencies, during
the main resource step. If a tunnel took a long time to establish, it'd
mean we'd start reporting every resource was taking a long time, which
is obviously no good!

This really exposes some of the jank which is PartialKey. I've got a WIP
branch which redoes the tunnel on top of a generic RPC system, which I
think will be much nicer.